update promo bar content
